stub out renderers for mock shortcode responses
Might serve as a starting point for real shortcode renderers that know
how to deal with the shortcode options.

Or it might just figure out how to pass it along to the old CERES
mechanisms to get the HTML then wrap it up for delivery here until
we get true BB integration.